+++ src/Framework/WordPress/Models/PostModel.php
@@ -181,24 +181,12 @@ class PostModel
     {
         $postId = $this->getPostId();
 
-        $metaKey = $this->hooks->applyFilters(
-            HooksAbstract::FILTER_ACTION_META_KEY,
-            $metaKey,
-            $postId
-        );
-
         return get_post_meta($postId, $metaKey, $single);
     }
 
     public function metadataExists($metaKey)
     {
         $postId = $this->getPostId();
-
-        $metaKey = $this->hooks->applyFilters(
-            HooksAbstract::FILTER_ACTION_META_KEY,
-            $metaKey,
-            $postId
-        );
 
         return metadata_exists('post', $postId, $metaKey);
     }